requirements.txt
async-timeout==4.0.2
Deprecated==1.2.14
packaging==23.1
python-dotenv==1.0.0
redis==4.3.4
redis-client==1.0.1
wrapt==1.15.0
